# apps/backend-api/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: TerraShare API
  version: "1.0.0"
  description: API para plataforma de alquiler de terrenos
paths:
...
  /lands:
    get:
      summary: List lands
      tags: [Lands]
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Full-text search query
        - name: lat
          in: query
          schema:
            type: number
          description: Latitude for geospatial search
        - name: lng
          in: query
          schema:
            type: number
          description: Longitude for geospatial search
        - name: radius
          in: query
          schema:
            type: number
            default: 10
          description: Search radius in km
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lands listed
    post:
      summary: Create a land
      tags: [Lands]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, area, location, priceRule, allowedUses]
              properties:
                title:
                  type: string
                description:
                  type: string
                area:
                  type: number
                allowedUses:
                  type: array
                  items:
                    type: string
                location:
                  type: object
                  properties:
                    province:
                      type: string
                    district:
                      type: string
                priceRule:
                  type: object
                  properties:
                    currency:
                      type: string
                      enum: [USD, PAB]
                    pricePerMonth:
                      type: number
      responses:
        "201":
          description: Land created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT